rename package from simulator to simulate
