+++ cmd/simv3/simulate/chain/listener_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/config"
-	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulator/chain"
+	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulate/chain"
 )
 
 func TestListener(t *testing.T) {
