+++ cmd/simv3/simulate/hydrator.go
@@ -1,14 +1,14 @@
-package simulator
+package simulate
 
 import (
 	"log"
 
 	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/config"
-	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulator/chain"
-	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulator/db"
-	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulator/net"
-	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulator/ocr"
-	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulator/upkeep"
+	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulate/chain"
+	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulate/db"
+	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulate/net"
+	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulate/ocr"
+	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulate/upkeep"
 	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/telemetry"
 	"github.com/smartcontractkit/ocr2keepers/pkg/v3/plugin"
 )
