+++ cmd/simv3/simulate/upkeep/active_test.go
@@ -8,8 +8,8 @@ import (
 	"time"
 
 	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/config"
-	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulator/chain"
-	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulator/upkeep"
+	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulate/chain"
+	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulate/upkeep"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
