+++ cmd/simv3/simulate/upkeep/perform.go
@@ -6,7 +6,7 @@ import (
 	"runtime"
 	"sync"
 
-	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulator/chain"
+	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulate/chain"
 	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/util"
 	ocr2keepers "github.com/smartcontractkit/ocr2keepers/pkg/v3/types"
 )
