+++ cmd/simv3/simulate/upkeep/source.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"math/big"
 
-	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulator/chain"
+	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/simulate/chain"
 	"github.com/smartcontractkit/ocr2keepers/cmd/simv3/util"
 	ocr2keepers "github.com/smartcontractkit/ocr2keepers/pkg/v3/types"
 )
